This was supposed to be a script to populate customer data with the correct badge numbers (since they currenlty just have *a* badge number for each SP, not necessarily the *correct* badge number).
However, this is impossible to do without making the customer data non-unique, and we really don't care about badge number per se, beyond its ability to link to pump test data.

So this script outputs a crosswalk at the SP-badge-date level. In other words, for any pump test at a given point time, we can map the badge number of the pump test to the correct SP/SA.