reworked things AGAIN to cycle through by month instead of by rate (because this lets me pare down the hourly interval data in a temp file BEFORE merging, which makes for a much faster merge)
also made the decision to outer-join by rate-group, for the choose-your-own-adventure rates. for now, we will assign each SA to the group with the highest within-SA correlation between bill$_actual and bill$_constructed (for the duration of the SA-rate). down the line, we can bring in group assignments before this outer-join...